refactor: fixing code climate code smell
